requirements.txt
tqdm
requests
requests_oauthlib
BeautifulSoup4
lxml
Flask
jupyter
pylint
python-dotenv
google-cloud-datastore
google-cloud-storage
google-cloud-logging
gunicorn
six
psq
ipython